cleanup printouts and unnecessary files
